Fixed import issues with dmidecode
As we now include libxml2 and the required libxml2mod (which is used to wrap
xmlDoc and xmlNode data into Python objects), importing only dmidecode caused
a failure.  If adding import libxml2 first, everything would work out fine.

To avoid this issue, due to backwards compatibility, a tiny dmidecode wrapper
is implemted as well.  The dmidecode.so module is now renamed to dmidecodemodule.so,
and the wrapper is now called dmidecode.py.

To simplify things, the dmidecodeXML module introduced in commit b25b2ca548508cd2beb26f465b7bc5a296592461
is not merged into the new dmidecode.py

The constants mentioned are now called dmidecode.DMIXML_NODE and dmidecode.DMIXML_DOC
and to instantiate the dmidecodeXML class, dmidecode.dmidecodeXML() is used.