Discovered another issue with Python and imports
The shard library got renamed to dmidecodemodule.so, and this was not
clever.  When you do 'import dmidecode' in Python, it will look for
files in this order:

	dmidecode
	dmidecode.so
	dmidecodemodule.so
	dmidecode.py
	dmidecode.pyc

This is of course a problem when the wrapper introduced in commit
65c9384ec9b6e265aba11227ffa37ae7a6a787d1 is called dmidecode.py, and
Python attempts to load dmidecodemodule.so before dmidecode.py.

To solve this, dmidecodemodule.so is now renamed to dmidecodemod.so.