Fixed wrong behavivour in pythonizeXMLnode() when no key value is found
If the XPath expression for the key value points on a non-existing XML
node, it would earlier abort pythonizing the XML data.  Even though this
could look like the correct behaviour, it will not work out well in
reality.

For sections like 'bios', it might be that the DMI/SMBIOS data do not
describe or provide the BIOSLanguage section (type id 0x0D).  Which
means calling dmidecode.bios() would fail completely instead of
reporting what it could find.

This patch fixes this issue and it will normally ignore not found key
values and just continue pythonizing the XML data and return what it
managed to translate into a Python dictionary.

If DEBUG is defined when compiling python-dmidecode, a warning message
will be printed directly to stderr when a key value is not found.

A warning is printed if dmidecode is compiled with DEBUG defined.