Added new function: dmixml_FindNodeByAttr_NoCase(...)
This function will ignore the case of the string of the value to be
found.  This improves the behaviour mentioned in
commit 20030e42b4d3f7283f6143641cb009a8dbf1da24.

At the moment, the immediate advantage is that the pymap.xml is
not strictly bound to if the type IDs in hex are in upper and/or
lower case.  Both cases and mixed cases will now work.