Added dmidump utility
This is a simple utility which dumps DMI/SMBIOS data straight
to a file, written i C and should not depend on anything extra.

Useful to get DMI/SMBIOS data without installing python and/or
python-dmidecode.  The data dump can be parsed by python-dmidecode
later on on a different box.