Improved set_dev() function
This function would often fail during the unit test.  Most probably because the
complete error checking was a bit odd.  Rewritten the error checks to be a bit
more sane.

Also improved error situations, where it will now throw an exception with an
error message instead of just returning False.