Updated examples/dmidump.py to be more up-to-date against a lot of new features
Rewrote bigger parts to be more intelligent, checking for warnings on the way.  Added
more warnings if script is run as a non-root user, and made it work as non-root user
if a dmidata.dump file exists (created when run as root).

Added examples with the XML API as well.