Add SIGILL catcher in mem_chunk()
On some s390x systems under heavy memory load seems to trigger
a SIGILL event when calling memcpy().  To avoid this issue, which
basically leads to the complete script to core dump, a signal
handler was implemented to add tackle such siutations more gracefully.

The SIGILL event should now be logged as well, to be able to see
that this happened.

Signed-off-by: David Sommerseth <dazo@users.sourceforge.net>