Do not add explictly 'dmispec' attributes inside switch() in dmi_decode()
The dmispec attribute is added outside the switch() call, and must not be
duplicated.  If this happens, an invalid XML file will be generated.
(Un)fortunately, libxml2 is quite forgiving to this error.  But xmllint
will complain about it and other XML libraries (such as python-lxml)
may reject such XML data.

Signed-off-by: David Sommerseth <davids@redhat.com>