Fix typo in pymap.xml - "Relase Date" -> "Release Date"
Thanks to Daniel Mueller for spotting this typo.

Unfortunately this fix may break some code if it depends on the
"Relase Date" key name.  The alternative would be to provide
both the old and wrong key name together with the new one, but
this particular key was not considered critically enough for special
treatment.

Signed-off-by: David Sommerseth <davids@redhat.com>